Allow override of params with filters
Although the comments in code say that setting the values  of "per_page" or "page" is possible using the "filters" parameter of the all() or search() methods, in reality those values are overwritten with "params.update()". Changed the behaviour for these functions to check whether the values to be updated in the "params" exist in filters. Update the values only if they are not present in filters.